* Patch by qknight adding a texcompress variant that works without   creating an OGL window. (closes mantis #636)
git-svn-id: http://springrts.com/svn/spring/trunk/tools/MapConv@4988 37977431-3df6-0310-b722-df95706aa16b

+++ FeatureCreator.cpp
@@ -1,10 +1,10 @@
 #include "FeatureCreator.h"
 #include "Bitmap.h"
+#include "math.h"
 #define NUM_TREE_TYPES 16
 #define treetop 215 /* TODO Shouldn't this be related to the previous define somehow? */
 extern float* heightmap;
 
-#include <math.h>
 #include <iostream> /* cout */
 using namespace std; /* cout */
 
@@ -208,8 +208,8 @@ void CFeatureCreator::PlaceVent(int x, int y, CBitmap * feature, CBitmap * vent,
 			x=bx+rand()*(40)/RAND_MAX-20;
 			y=by+rand()*(40)/RAND_MAX-20;
 #else
-			x=(int)(bx + ((float)(rand())*(40.0) )/RAND_MAX-20);
-			y=(int)(by + ((float)(rand())*(40.0) )/RAND_MAX-20);
+			x=bx+( (float)(rand())*(40.0) )/RAND_MAX-20;
+			y=by+( (float)(rand())*(40.0) )/RAND_MAX-20;
 #endif
 			if(x<5)
 				x=5;
